nix: provide cabal and ghc
Define a minimal development environment providing cabal-install and
ghc. Note that `import <nixpkgs>` references the nixpkgs found in the
user's NIX_PATH, which can be shown by running:

  echo $NIX_PATH | sed 's/:/\n/g' | grep '^nixpkgs=' | cut -d= -f2

This is sufficient for demonstration purposes, but in order to ensure
reproducible builds, we'll want to specify exact versions, especially
for ghc, rather than live at the mercy of whichever versions are present
in the user's nix profile.